FileHandler.java: JavaDoc, improved logging, dir checks

+++ src/main/java/evergoodteam/chassis/util/handlers/FileHandler.java
@@ -10,75 +10,70 @@ import java.io.IOException;
 import java.io.Writer;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 
-import static evergoodteam.chassis.util.Reference.MODID;
+import static evergoodteam.chassis.util.Reference.CMI;
 import static org.slf4j.LoggerFactory.getLogger;
 
 public class FileHandler {
 
-    private static final Logger LOGGER = getLogger(MODID + "/H/File");
+    private static final Logger LOGGER = getLogger(CMI + "/H/File");
 
     public static void emptyFile(Path path) {
         emptyFile(path.toFile());
     }
 
     public static void emptyFile(File file) {
-
         if (file.exists()) {
             try {
                 new FileWriter(file, false).close();
+                //LOGGER.info("Emptied file at {}", file.getPath());
             } catch (IOException e) {
-                LOGGER.error("Error on emptying File at {}", file.getPath(), e);
+                LOGGER.error("Error on emptying file at {}", file.getPath(), e);
             }
         } else {
-            LOGGER.warn("Provided Path is not valid: {}", file.getPath());
+            LOGGER.warn("Provided path is not valid: {}", file.getPath());
         }
     }
 
     /**
-     * Writes provided String to the File identified with the Path
-     *
-     * @param text
-     * @param path
+     * Writes the provided string to the file at the specified path
      */
     public static void writeToFile(String text, Path path) {
-
         //LOGGER.info("Attempting to write to path {}", path);
-
         if (Files.exists(path)) {
             try (Writer fileWriter = new FileWriter(path.toFile())) {
                 fileWriter.write(text);
             } catch (IOException e) {
-                LOGGER.error("Error on writing to File at {}", path, e);
+                LOGGER.error("Error on writing to file at {}", path, e);
             }
         } else {
-            LOGGER.warn("Provided Path is not valid: {}", path);
+            LOGGER.warn("Provided path is not valid: {}", path);
         }
     }
 
+    /**
+     * Creates a json file at the specified path
+     *
+     * @param path NOTE: ".json" extension is not required
+     */
     public static void createJsonFile(Path path) {
-        Path actual = StringUtils.checkDuplicateExtension(path, ".json");
-        //LOGGER.info(actual);
-        createFile(actual, ".json");
+        createFile(path, ".json");
     }
 
     /**
-     * Creates a File only if one with the same Path doesn't exist already
+     * Creates a file only if one with the same Path doesn't exist already
      *
-     * @param path
-     * @param extension .txt .json etc.
+     * @param path      file path
+     * @param extension e.g. ".txt", ".json"
      */
     public static void createFile(Path path, String extension) {
-        Path actual = StringUtils.checkDuplicateExtension(path, extension);
-        createFile(Paths.get(actual + extension));
+        createFile(StringUtils.addExtension(path, extension));
     }
 
     /**
-     * Creates a File only if one with the same Path doesn't exist already
+     * Creates a file only if one with the same path doesn't exist already
      *
-     * @param path Note: must include the File extension
-     * @throws IOException
+     * @param path NOTE: must include the file extension
      */
     public static void createFile(Path path) {
         if (!Files.exists(path)) {
@@ -92,19 +87,18 @@ public class FileHandler {
     }
 
     /**
-     * Delete a File at the specified Path
+     * Deletes a file at the specified path
      *
-     * @param path path of your File
-     * @throws java.nio.file.DirectoryNotEmptyException if used with Dirs
+     * @param path path of your file
      */
     public static void delete(Path path) {
         try {
-            if (Files.exists(path)) {
+            if (Files.exists(path) && !Files.isDirectory(path)) {
                 FileUtils.delete(path.toFile());
-                //LOGGER.info("Deleted File at {}", path);
+                //LOGGER.info("Deleted file at {}", path);
             }
         } catch (IOException e) {
-            LOGGER.error("Error on deleting File at {}", path, e);
+            LOGGER.error("Error on deleting file at {}", path, e);
         }
     }
 }
